web/requirements.txt
fastapi ~= 0.66.0 
uvicorn ~= 0.14
python-dotenv ~= 0.18.0
pydantic ~= 1.8.2 
pytest ~= 6.2.4
pylint ~= 2.12
python-jose[cryptography] ~= 3.3
httpx ~= 0.21
python-multipart ~= 0.0.5 
openapi-python-client ~= 0.10
